zerobus-snap-demo/pyproject.toml
[project]
name = "zerobus-snap-demo"
version = "0.2.0"
description = "Zerobus Ingest Demo — SecondDinner / Marvel Snap Edition"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiohttp>=3.9.0",
    "databricks-sdk>=0.30.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
    "databricks-zerobus-ingest-sdk>=1.0.0",
]
